Extra function can be a permanent part of theme
The function that the instructions tell you to add to make this work can now be added to all themes. The plugin will then automatically be activated when the plugin is installed. As an added bonus, should the plugin be removed down the road (not wanted, no longer supported, etc), it won't break the theme because someone forgot to remove the extra function from their lib.php.